requirements.txt
requests
python-decouple
pydantic 
aiohttp 
psycopg2-binary